Release testing image to arle prod (#460)
* Debian 11 prod publish task

Previously there were no tasks in the arle testing pipeline to test an image publish without publishing to a public cloud project.

Now referencing a static debian-11 buid we can use this entry to test in production.

* Moved workflow to compute-image-tools

Also made the prod test generic so that an accidental release is never done.

* Added testing image to arle-prod in its own job