Update Windows pipeline to include new environment (#562)
* Update Windows pipeline to include new environment

Adds "client" environment to builds; removes excess publish job as I don't think it's needed anymore.

* Revert to 2 publish jobs for clients

Just kidding, we still need it.

* Update environments for client builds

Working around not having prod with the current pipeline is tricky. This should resolve it.

* Fixing client_envs values

Last change broke image groupings. I'm not happy with how this is set up now, but it'll work.

* Add client to ImgPublishJob

Missed adding this resulting in blank job name

* Fix publish name in passed criteria for client env

Typos are hard