Update wait-for-build timeout for package builds (#580)
* Update timeout thresholds for package builds

Default timeout of 10m has recently become too short for some builds of EL; increasing to 20m to allow proper cleanup.

* Move timeout definition for package build steps

I believe this is the correct definition location for timeouts as the previous definition was being superceded by the default 10m timeout on the parent task.

* Update wait-for-build timeout for package builds

Fixes the 10m timeout superceding the previous 20m adjustment.