Fork windows image build for staging builds (#594)
* Fork windows image build for staging builds

Enables build and publish to testing for images using the staging Rapture repo. Stripped all publishing tasks to prod and internal for safety.

* Update job and env variables in staging pipeline

Tweaking to adjust for removal of other envs

* Update job names

Missed a few references to the old env schema

* Update publish names

Mistakenly used testing instead of staging to distinguish builds.

* Add staging suffix to img group defs

Missed it in the groupings

* Update pipeline names to staging; task names to testing

Reflects publish environments and distinctions from prod pipeline

* Update task name staging --> testing

Missed one in the last update.

* Add Windows staging build to pipelines

This is needed to render the json and set the pipeline config with pipeline-set-pipeline