Add the ability for TestVM to run on an instance from the beta API.
The abstraction of setting one of instance or instancebeta is not super elegant but it matches the abstractions present in daisy. There is no unified way to represent a compute instance which is either GA or beta. Generics are not an option because we can't directly reference arbitrary properties on instiated generic types, something we need to do frequently.

Added support for beta instances in existing TestVM methods. Expanded unit tests for TestVM methods to cover instancebeta cases.
I modified the behavior in a couple TestVM methods which were clobbering ShieldedInstanceConfig, it's now possible to enable both secure boot and shielded vm in the same instance without them clobbering each other.

Added a function to create a TestVM from a daisy.InstanceBeta and a slice of compute (GA) disks. Added unit tests for this.