Add LSSD case for hotattach
This isn't actually attaching, but the rest of the test behaves the same which is why I stuck it in this suite.

I reworked the hotattach utility functions slightly to make it a bit more flexible. The disk name is passed to the guest through the MDS now.

NVMe disks surface the disk name through friendlyname so I used that to identify disks when available, falling back to just using disk 1 when not.