Add a prep command for guest agent reinstallation (#1025)
Necessary to avoid attempting to reinstall a pkg removed from repos on distros with apt.

This was getting to be a lot of duplicated logic so I abstracted the logic into a general case. I considered putting this until test_utils but no other test suite needs this AFAIK so I left it in metadata.