Add LSSD support and cases to storageperf LSSDs present on the machine are collected into a raid array as directed in our public documentation
I abstracted installFioLinux to install any package and used that to install mdadm for the array setup, then added some logic to isntall libnuma-dev while I was at it

Added performance targets for LSSDs

Added some logic to determine "disk class" eg. PD, hyperdisk, or LSSD and reworked logic to return fio options based off that

getMountDiskSize returns 0 for LSSD so that it doesn't wait for quota it doesn't need

Test setup logic just checks of diskType is lssd and skips adding a secondary disk if so.