Refactor licensevalidation suite. (#1060)
* Refactor licensevalidation suite.
Instead of checking a hard-coded list of license codes, generate a list of expected licenses based on what we know about the image and compare that to the license urls on it.
Then check that all license codes present on the image propogate to the VM created from it.

While I was messing with these suite I internationalized the ActivationCheck per the FR for that

* golint

* gofmt