Update CIT to run only during testing publish for Windows (#1167)
* Update CIT to run only during testing publish

Probably to be moved to prod when we can fully automate the publish process. For now, we don't want to run the tests on publish-to-prod as thats the final human step to send it to ARLE, and we should be blocking on publish-to-testing if CIT breaks.

* Update windows-image-build.jsonnet

* Revert staging CIT logic

Env doesn't exist here, so this should be left as is.

* Update runtests to exclude client