container-images: install protobuffer tooling.
In order to be able to generate pb.go files during build time we need
to have installed the protobuffer compiler and the go-grpc generator.

google-guest-agent repository is not checking in generated code but
relying on its build system to generate it during build time, prow
nodes must have the required tooling to do so.