package build: selectively pass daisy args & pipeline shift (#1258)
Concourse will try to interpolate variables even if they are not set
in that case if the variable is not set concourse will fail with an
"no such variable" kind of error - this change fixes that.

Additionally originally we set extra_repo to the wrong build pipeline,
we set it to guest-agent-dev when we should have set it to guest-agent.