rpmbuild: properly set root dir and enable tracing
As git_checkout() is not popping from the checkout dir we were setting
the root directory inside the package's source code (not the expected),
for that then we set the variable before the first git_checkout() call.

Additionally we are setting -e and -x flags so the it fails early and
also prints the script tracing.