Disable accelerator CIT tests on linux builds and split accelerator images into Rocky/Ubuntu (#1293)
* Disable accelerator CIT tests on linux builds

Cannot run until NIC types rolls out.

* Test passing in array for rocky images

Not sure if this works

* Fix syntax

The brackets may not have been necessary

* Split rocky and ubuntu image groups

Allows for easier separation and removal of ubuntu when the time comes; we wont actually build and publish it.

* Remove extra comma

* Fix comma

Was missing one, not an extra

* Fixing rocky reference

* Update incorrect reference

* Remove ubuntu from accelerator builds

* Rename accelerator_images

Specify rocky to leave it open for others

* Fix work_flow dir

* Move rocky accelerator sbom under other rocky sboms; fix directory