Use Debian 12 worker in Deb13/RHEL10 package build
The default is currently still set to Debian 11 for the worker images. We should update this in the future, but for now let's be explicit in using the Debian 12 worker to get things running again.

I plan to test using Debian 12 as a default in all builds after discovering this, but let's not potentially break existing builds by moving workers.