Add support to target specific Debian versions in package build (#1506)
* Add support to target specific Debian versions in package build

This adds an optional variable that when set, can be used by the debhelper to build for specific Debian versions, but retain the previous behavior of building for "all" when it is not set.

* Only pass in TARGET_VERSION when guest-configs is built

This passes in the variable only when guest-configs is built. Nothing else uses it for now, so this makes sure we're not passing  variables in needlessly.