Remove package upgrade and windows containers from windows build tests (#1559)
* Remove package upgrade from windows build tests

These specifically test the -testing repositories, which shouldn't block builds that are using the -stable repositories.

* Remove sql and container tests

These do not apply to the images we're building in this pipeline

* Put sql back into test filter

I take back the last comment, this is needed for the sql derivative builds unless we want to split out the templates.

* Restore original order

Small detail but easier to diff